+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+Initial Version of the library.
+
+    * Includes Form Component
+    * Includes Form Checkbox Component
+    * Includes Form Datepicker Component
+    * Includes Form Dropdown Component